Factor out common xcode version test utility.
